+++ BUILD.gn
@@ -66,6 +66,7 @@ ts_sources = [
   "js/libdeno.ts",
   "js/main.ts",
   "js/mkdir.ts",
+  "js/make_temp_dir.ts",
   "js/mock_builtin.js",
   "js/os.ts",
   "js/plugins.d.ts",
