+++ js/deno.ts
@@ -1,8 +1,9 @@
 // Copyright 2018 the Deno authors. All rights reserved. MIT license.
 // Public deno module.
 /// <amd-module name="deno"/>
-export { env, exit, makeTempDirSync } from "./os";
+export { env, exit } from "./os";
 export { mkdirSync, mkdir } from "./mkdir";
+export { makeTempDirSync, makeTempDir } from "./make_temp_dir";
 export { removeSync, remove, removeAllSync, removeAll } from "./remove";
 export { readFileSync, readFile } from "./read_file";
 export { renameSync, rename } from "./rename";
