+++ js/os_test.ts
@@ -22,41 +22,3 @@ test(async function envFailure() {
 
   assert(caughtError);
 });
-
-testPerm({ write: true }, function makeTempDirSync() {
-  const dir1 = deno.makeTempDirSync({ prefix: "hello", suffix: "world" });
-  const dir2 = deno.makeTempDirSync({ prefix: "hello", suffix: "world" });
-  // Check that both dirs are different.
-  assert(dir1 != dir2);
-  for (const dir of [dir1, dir2]) {
-    // Check that the prefix and suffix are applied.
-    const lastPart = dir.replace(/^.*[\\\/]/, "");
-    assert(lastPart.startsWith("hello"));
-    assert(lastPart.endsWith("world"));
-  }
-  // Check that the `dir` option works.
-  const dir3 = deno.makeTempDirSync({ dir: dir1 });
-  assert(dir3.startsWith(dir1));
-  assert(/^[\\\/]/.test(dir3.slice(dir1.length)));
-  // Check that creating a temp dir inside a nonexisting directory fails.
-  let err;
-  try {
-    deno.makeTempDirSync({ dir: "/baddir" });
-  } catch (err_) {
-    err = err_;
-  }
-  assertEqual(err.kind, deno.ErrorKind.NotFound);
-  assertEqual(err.name, "NotFound");
-});
-
-test(function makeTempDirSyncPerm() {
-  // makeTempDirSync should require write permissions (for now).
-  let err;
-  try {
-    deno.makeTempDirSync({ dir: "/baddir" });
-  } catch (err_) {
-    err = err_;
-  }
-  assertEqual(err.kind, deno.ErrorKind.PermissionDenied);
-  assertEqual(err.name, "PermissionDenied");
-});
